:recycle: refactor: Refatorando renderização do componente de CardPauta
